remove python2 + rename package
